server/requirements_comfy.txt
torch>=2.0.0
torchvision
diffusers>=0.24.0
transformers>=4.25.0
accelerate>=0.20.0
safetensors>=0.3.0
Pillow>=9.0.0
flask>=2.3.0
flask-cors>=4.0.0
xformers